uefi::helper: clean up, multiple sub modules
I didn't change anything about the implementations, just moved the existing code into the
corresponding modules.
