+++ uefi/src/allocator.rs
@@ -130,5 +130,3 @@ unsafe impl GlobalAlloc for Allocator {
         (*boot_services()).free_pool(ptr).unwrap();
     }
 }
-
-
