+++ uefi/src/helpers/global_allocator.rs
@@ -1,0 +1,4 @@
+use crate::allocator::Allocator;
+
+#[global_allocator]
+static ALLOCATOR: Allocator = Allocator;
