+++ uefi/src/lib.rs
@@ -127,7 +127,6 @@ pub mod helpers;
 
 mod util;
 
-
 #[cfg(test)]
 // Crates that create procedural macros can't unit test the macros they export.
 // Therefore, we do some tests here.
